Refactor and extend to support more --extract options
The --extract option now accepts a bunch more options to pull out the
separate types of tweets. This is just a convenience, to save people
having to filter the output to extract different types.

Along with that, I refactored the internal code architecture a little
bit to try to keep things understandable and maintainable. I hope.